+++ drizzle/0024_clumsy_xavin.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "eve_npc_stations" ADD COLUMN "name" text;